Move projectfile into subpackage refs #219
