+++ pyrs/core/nexus_conversion.py
@@ -9,9 +9,8 @@ import numpy
 import os
 from pyrs.core import workspaces
 from pyrs.core.instrument_geometry import AnglerCameraDetectorGeometry, HidraSetup
-from pyrs.projectfile import HidraConstants
+from pyrs.projectfile import HidraConstants, HidraProjectFile, HidraProjectFileMode
 from pyrs.utilities import checkdatatypes
-from pyrs.utilities.rs_project_file import HidraProjectFile, HidraProjectFileMode
 
 
 class NeXusConvertingApp(object):
