+++ scripts/preparetest/convert_strain_stress_data.py
@@ -2,8 +2,7 @@
 # System (library) test for classes and methods that will be used to calculate strain and stress
 import os
 import sys
-from pyrs.projectfile import HidraProjectFileMode
-from pyrs.utilities.rs_project_file import HidraProjectFile
+from pyrs.projectfile import HidraProjectFile, HidraProjectFileMode
 
 
 def contain_strain_stress_main():
