+++ tests/integration/test_calibration_class.py
@@ -5,9 +5,8 @@ import pytest
 import time
 import os
 import json
-from pyrs.projectfile import HidraProjectFileMode
+from pyrs.projectfile import HidraProjectFile, HidraProjectFileMode
 from pyrs.utilities import calibration_file_io
-from pyrs.utilities.rs_project_file import HidraProjectFile
 try:
     from pyrs.calibration import peakfit_calibration
 except ImportError as e:
