+++ tests/integration/test_peak_fitting.py
@@ -6,7 +6,7 @@ from pyrs.peaks import PeakCollection
 from pyrs.core.peak_profile_utility import PeakShape, BackgroundFunction
 from pyrs.core.summary_generator import SummaryGenerator
 from pyrs.dataobjects import SampleLogs
-from pyrs.utilities.rs_project_file import HidraProjectFile
+from pyrs.projectfile import HidraProjectFile
 from pyrs.peaks import PeakFitEngineFactory
 import h5py
 from pyrs.core import peak_profile_utility
